Static analysis warnings

+++ far/interf.cpp
@@ -725,7 +725,7 @@ void SetCursorType(bool const Visible, size_t Size)
 	if (Size == static_cast<size_t>(-1) || !Visible)
 	{
 		const size_t index = IsConsoleFullscreen()? 1 : 0;
-		Size = Global->Opt->CursorSize[index]? static_cast<int>(Global->Opt->CursorSize[index]) : InitialCursorInfo.dwSize;
+		Size = Global->Opt->CursorSize[index]? static_cast<size_t>(Global->Opt->CursorSize[index]) : InitialCursorInfo.dwSize;
 	}
 	Global->ScrBuf->SetCursorType(Visible, Size);
 }
